# docs/swagger.yaml
openapi: 3.0.3
info:
  title: WeAllBeam API
  description: |
    Educational management system API for schools, teachers, students, and parents.

    ## Language Support
    API supports multilingual content. Use language prefix in URLs:
    - English: `/en/endpoint`
    - Arabic: `/ar/endpoint`
  version: "5.10.23"
servers:
  - url: http://localhost:4000/ar
    description: Development server (arabic)
  - url: http://localhost:4000/en
    description: Development server (english)
  - url: https://api.weallbeam.com/ar
    description: Production server (arabic)
  - url: https://api.weallbeam.com/en
    description: Production server (english)
security:
  - bearerAuth: []
...
paths:
...
  /profile:
    get:
      tags: [Profile]
      summary: Get user profile
      description: Get current user's profile information based on role
      responses:
        "200":
          description: Profile information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      tags: [Profile]
      summary: Update user profile
      description: Update current user's profile information
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                phone:
                  type: string
                photo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"

  # ==================== Super Admin ====================
  /super-admin:
    post:
      tags: [Super Admin]
      summary: Create super admin
      description: Create a new super admin account (super-admin only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateSuperAdminRequest"
      responses:
        "201":
          description: Super admin created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuperAdmin"
        "400":
          $ref: "#/components/responses/ValidationError"
  /super-admin/{id}:
    get:
      tags: [Super Admin]
      summary: Get super admin by ID
      description: Get super admin details
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Super admin details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuperAdmin"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      tags: [Super Admin]
      summary: Update super admin
      description: Update super admin personal information
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateSuperAdminRequest"
      responses:
        "200":
          description: Super admin updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuperAdmin"
...
  /school:
    post:
      tags: [Schools]
      summary: Create school
      description: Create a new school (super-admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolRequest"
      responses:
        "201":
          description: School created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "400":
          $ref: "#/components/responses/ValidationError"

    get:
      tags: [Schools]
      summary: List schools
      description: Get list of all schools (super-admin only)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Schools list
          content:
            application/json:
              schema:
                type: object
                properties:
                  schools:
                    type: array
                    items:
                      $ref: "#/components/schemas/School"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /school/{id}:
    get:
      tags: [Schools]
      summary: Get school by ID
      description: Get detailed information about a specific school
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: School details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      tags: [Schools]
      summary: Update school
      description: Update school information
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSchoolRequest"
      responses:
        "200":
          description: School updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"

    delete:
      tags: [Schools]
      summary: Delete school
      description: Delete a school and all related data
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: School deleted successfully
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ==================== Teachers ====================
  /teachers:
    post:
      tags: [Teachers]
      summary: Create teacher
      description: Add a new teacher (super-admin, school only)
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "201":
          description: Teacher created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "400":
          $ref: "#/components/responses/ValidationError"

    get:
      tags: [Teachers]
      summary: List teachers
      description: Get list of teachers (school can see their teachers)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Teachers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    get:
      tags: [Teachers]
      summary: Get teacher by ID
      description: Get teacher details
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Teacher details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      tags: [Teachers]
      summary: Update teacher
      description: Update teacher information
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateTeacherRequest"
      responses:
        "200":
          description: Teacher updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"

    delete:
      tags: [Teachers]
      summary: Delete teacher
      description: Delete a teacher
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Teacher deleted successfully
...
  /students:
    post:
      tags: [Students]
      summary: Create student
      description: Add a new student (super-admin, school, teacher only)
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "201":
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/ValidationError"

    get:
      tags: [Students]
      summary: List students
      description: Get list of students
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      tags: [Students]
      summary: Get student by ID
      description: Get student details
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      tags: [Students]
      summary: Update student
      description: Update student information
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateStudentRequest"
      responses:
        "200":
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

    delete:
      tags: [Students]
      summary: Delete student
      description: Delete a student
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Student deleted successfully
...
  /parent:
    post:
      tags: [Parents]
      summary: Create parent
      description: Add a new parent (super-admin, school, teacher only)
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateParentRequest"
      responses:
        "201":
          description: Parent created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"

    get:
      tags: [Parents]
      summary: List parents
      description: Get list of parents
      responses:
        "200":
          description: Parents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Parent"
  /parent/{id}:
    get:
      tags: [Parents]
      summary: Get parent by ID
      description: Get parent details
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Parent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      tags: [Parents]
      summary: Update parent
      description: Update parent information
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateParentRequest"
      responses:
        "200":
          description: Parent updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"

    delete:
      tags: [Parents]
      summary: Delete parent
      description: Delete a parent (super-admin, school only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Parent deleted successfully

  # ==================== Classes ====================
  /classes:
    post:
      tags: [Classes]
      summary: Create class
      description: Create a new class (super-admin, school only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Class created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"

    get:
      tags: [Classes]
      summary: List classes
      description: Get list of all classes
      responses:
        "200":
          description: Classes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
  /classes/{id}:
    get:
      tags: [Classes]
      summary: Get class by ID
      description: Get class details
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Class details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"

    put:
      tags: [Classes]
      summary: Update class
      description: Update class information
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassRequest"
      responses:
        "200":
          description: Class updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"

    delete:
      tags: [Classes]
      summary: Delete class
      description: Delete a class
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Class deleted successfully
...
  /questions:
    post:
      tags: [Questions]
      summary: Create form
      description: Create a new form with questions (super-admin, school only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFormRequest"
      responses:
        "201":
          description: Form created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"

    get:
      tags: [Questions]
      summary: List all forms
      description: Get list of all available forms
      responses:
        "200":
          description: Forms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Form"
...
  /questions/{subject}:
    get:
      tags: [Questions]
      summary: Get form by subject
      description: Get form by subject name
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
          example: daily
      responses:
        "200":
          description: Form details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"

    put:
      tags: [Questions]
      summary: Update form by subject
      description: Update form questions and settings
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFormRequest"
      responses:
        "200":
          description: Form updated successfully

    delete:
      tags: [Questions]
      summary: Delete form by subject
      description: Delete a form
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Form deleted successfully
...
  /responses/{id}:
    get:
      tags: [Responses]
      summary: Get school statistics
      description: Get response statistics for a school by date range
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: fromDay
          in: query
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: toDay
          in: query
          schema:
            type: string
            format: date
          example: "2024-01-31"
      responses:
        "200":
          description: School statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolStatistics"
...
  /blog:
    post:
      tags: [Blog]
      summary: Create blog post
      description: Create a new blog post (super-admin only)
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateBlogRequest"
      responses:
        "201":
          description: Blog post created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"

    get:
      tags: [Blog]
      summary: List blog posts
      description: Get paginated list of blog posts (public)
      security: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, private, both]
      responses:
        "200":
          description: Blog posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogListResponse"
...
  /blog/{id}:
    get:
      tags: [Blog]
      summary: Get blog post by ID
      description: Get detailed blog post information
      security: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Blog post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"

    put:
      tags: [Blog]
      summary: Update blog post
      description: Update existing blog post
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateBlogRequest"
      responses:
        "200":
          description: Blog post updated successfully

    delete:
      tags: [Blog]
      summary: Delete blog post
      description: Delete a blog post and associated files
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Blog post deleted successfully

  # ==================== PDF ====================
...
  /pdf:
    get:
      tags: [PDF]
      summary: List PDFs
      description: Get paginated list of PDFs
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          schema:
            type: string
        - name: targetSchools
          in: query
          schema:
            type: string
        - name: supportedLanguages
          in: query
          schema:
            type: string
      responses:
        "200":
          description: PDFs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PDFListResponse"
...
  /pdf/{id}:
    get:
      tags: [PDF]
      summary: Get PDF by ID
      description: Get PDF details (public view)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: PDF details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PDF"

    put:
      tags: [PDF]
      summary: Update PDF
      description: Update PDF metadata and files
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdatePDFRequest"
      responses:
        "200":
          description: PDF updated successfully

    delete:
      tags: [PDF]
      summary: Delete PDF
      description: Delete a PDF document
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: PDF deleted successfully
...
  /ticket/{id}:
    get:
      tags: [Tickets]
      summary: Get ticket by ID
      description: Get ticket details
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
  /incidents:
    post:
      tags: [Incidents]
      summary: Create incident
      description: Report a new incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentRequest"
      responses:
        "201":
          description: Incident created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
...
  /incidents/{incidentId}:
    get:
      tags: [Incidents]
      summary: Get incident by ID
      description: Get incident details
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

    put:
      tags: [Incidents]
      summary: Update incident
      description: Update incident information
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentRequest"
      responses:
        "200":
          description: Incident updated successfully

    delete:
      tags: [Incidents]
      summary: Delete incident
      description: Delete an incident
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident deleted successfully

  # ==================== Forum ====================
  /forum:
    post:
      tags: [Forum]
      summary: Create forum post
      description: Create a new forum post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateForumPostRequest"
      responses:
        "201":
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPost"

    get:
      tags: [Forum]
      summary: List forum posts
      description: Get all forum posts
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForumPost"
  /forum/{id}:
    get:
      tags: [Forum]
      summary: Get post with replies
      description: Get forum post with all replies
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Post with replies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPostWithReplies"
...
  /faqs:
    post:
      tags: [FAQ]
      summary: Create FAQ
      description: Create a new FAQ entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFAQRequest"
      responses:
        "201":
          description: FAQ created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FAQ"

    get:
      tags: [FAQ]
      summary: List FAQs
      description: Get all FAQs
      security: []
      responses:
        "200":
          description: FAQs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FAQ"
  /faqs/{id}:
    put:
      tags: [FAQ]
      summary: Update FAQ
      description: Update FAQ entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFAQRequest"
      responses:
        "200":
          description: FAQ updated successfully

    delete:
      tags: [FAQ]
      summary: Delete FAQ
      description: Delete FAQ entry
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: FAQ deleted successfully

  # ==================== Contacts ====================
  /contacts:
    post:
      tags: [Contacts]
      summary: Create contact message
      description: Submit a contact message
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

    get:
      tags: [Contacts]
      summary: List contact messages
      description: Get all contact messages (super-admin only)
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"

  # ==================== Filter (Data) ====================
...
  /filter/classes:
    get:
      tags: [Filter]
      summary: Get classes for filter
      description: Get classes list for dropdown filters
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Classes list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    name:
                      type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
...
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
...
  schemas:
...
    SuperAdmin:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        photo:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateSuperAdminRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
...
    ProfileResponse:
      type: object
      properties:
        _id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        photo:
          type: string
        role:
          type: string
        school:
          type: object
          nullable: true
        class:
          type: object
          nullable: true

    # ==================== Common ====================
...